Set 0600 permissions on test_credentials pre-test
The RSpec tests utilize the `test_credentials` file as a sample
credential file. Git does not manage granular file permissions,
only the execute and symlink bits. So in order to give this file
the 0600 permissions that rubygems wants it to have, this patch
forces that permission prior to running the tests.