Fix broken tests
- the overloaded `load_api_keys` in `lib/keycutter/configuration.rb`
was using a local variable named `rubygems_api_key` instead of an
instance variable, so the "active" `@rubygems_api_key` was never
getting set. This commit changes the local var to the instance var.
The base rubygems version of this method does exactly this, so
it's definitely the right place for it to be set.

- `Gem.configuration.load_rubygems_api_key` was giving a
"NoMethodError". The `@rubygems_api_key` is now set when calling
`load_api_keys`, so this call has been changed to `load_api_keys`.

- Defaulting the `@api_keys` to the value of `@hash` when there is
no `~/.gem/credentials` file is problematic, because that file could
contain values that do not belong in the credentials file. It does
not appear that keycutter ever properly worked in this scenario, so
this commit changes the default value to an empty hash to prevent
strange behavior when the credentials file does not exist.

- Rubygems now requires that the credentials file have 0600
permissions set to operate against it. This commit sets the
permissions mask in the `File.open()` call so that in the case when
the file did not already exist, it is created with the correct
permissions.

Fixes #3. Fixes #4. Fixes #5.